Declare topology and init counter before starting endpoints
